+++ src/components/Footer.vue
@@ -1,9 +1,25 @@
 <template>
   <footer class="footer has-margin-top-30">
     <div class="columns">
-      <div class="column">
-        <g-link to="/recommendations" class="is-size-7">Recommendations</g-link>
+      <div class="column has-text-centered">
+        Made with
+        &nbsp;
+        <b-icon pack="fas" icon="heart" size="is-small" class="red" />&nbsp;
+        on
+        &nbsp;
+        <a
+          href="https://github.com/ngunyimacharia/ngunyimacharia/"
+          target="_blank"
+        >
+          <b-icon pack="fab" icon="github" size="is-small" />
+        </a>
       </div>
     </div>
   </footer>
 </template>
+
+<style scoped>
+.red {
+  color: red;
+}
+</style>