+++ changelog.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Recommendations complete
 - Google sheets connection complete
 - Hasura source added and country to form
+- Jokes source added and configured
 
 ## [0.0.0] - 2019-01-14